Extract scalar tests into separate file
